refactor: organize auth-service packages
