+++ auth-service/src/main/java/morning/com/services/auth/dto/AuthRequest.java
@@ -1,4 +1,4 @@
-package morning.com.services.auth.model;
+package morning.com.services.auth.dto;
 
 public record AuthRequest(String username, String password) {
 }
