+++ auth-service/src/main/java/morning/com/services/auth/dto/AuthResponse.java
@@ -1,4 +1,4 @@
-package morning.com.services.auth.model;
+package morning.com.services.auth.dto;
 
 public record AuthResponse(String token) {
 }
