+++ auth-service/src/main/java/morning/com/services/auth/dto/ResultEnum.java
@@ -1,4 +1,4 @@
-package morning.com.services.auth.model;
+package morning.com.services.auth.dto;
 
 /**
  * Standard result codes and messages for API responses.
