+++ auth-service/src/main/java/morning/com/services/auth/entity/User.java
@@ -1,4 +1,4 @@
-package morning.com.services.auth.model;
+package morning.com.services.auth.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
